requirements.txt
streamlit==0.84.0
plotly==5.1.0
tensorflow-cpu==2.5.0
Pillow==9.5.0
